CodeGen from PR 14172 in Azure/azure-rest-api-specs Merge b918240a3393166410e1c8cb6be3368865bc7322 into e7682aa897902920f3a95b2f358b6f7729d18666

+++ services/preview/cognitiveservices/v3.0/luis/authoring/models.go
@@ -706,7 +706,7 @@ type LabelExampleResponse struct {
 	// UtteranceText - The example utterance.
 	UtteranceText *string `json:"UtteranceText,omitempty"`
 	// ExampleID - The newly created sample ID.
-	ExampleID *int32 `json:"ExampleId,omitempty"`
+	ExampleID *int64 `json:"ExampleId,omitempty"`
 }
 
 // LabelTextObject an object containing the example utterance's text.
